# specs/grc-claw-openapi.yaml
openapi: 3.1.0
info:
  title: GRC_Claw API
  description: |
    GRC_Claw — open-source governance, risk, and compliance platform for agentic AI.
    
    This API provides four complementary interfaces:
    - **REST** — CRUD operations, search, and reporting
    - **GraphQL** — Complex nested queries for dashboards and analytics
    - **gRPC** — High-performance enforcement decisions and streaming
    - **Webhooks** — Event-driven integrations
    
    ## Authentication
    
    All API requests require authentication via one of:
    - **OAuth 2.1 / OIDC** — Bearer JWT token (RS256)
    - **API Key** — `Authorization: Bearer grc_live_...` or `grc_test_...`
    - **mTLS (SPIFFE)** — X.509 SVID for agent-to-PEP communication
    
    ## Rate Limiting
    
    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit` — Request limit per window
    - `X-RateLimit-Remaining` — Remaining requests in current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets
    
    ## Error Format
    
    All errors follow RFC 7807 (Problem Details for HTTP APIs).
    
    ## Versioning
    
    Current version: `v1.0`. Version is specified in the URL path.
    Deprecation notices are sent via webhook 6 months before sunset.
  version: 1.0.0
  contact:
    name: GRC_Claw Architecture Team
    url: https://api.grc-claw.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.grc-claw.io
...
paths:
...
  /v1.0/evidence:
    get:
      operationId: searchEvidence
      summary: Search evidence
      description: Search evidence with filtering and full-text search.
      tags: [Evidence]
      parameters:
        - name: policy_id
          in: query
          schema: { type: string }
        - name: assessment_id
          in: query
          schema: { type: string }
        - name: evidence_type
          in: query
          schema:
            type: string
            enum: [artifact, observation, interview, analysis, log]
        - name: framework
          in: query
          schema: { type: string }
        - name: control_id
          in: query
          schema: { type: string }
        - name: verification_level
          in: query
          schema:
            type: string
            enum: [L0, L1, L2, L3, L4]
        - name: environment
          in: query
          schema: { type: string }
        - name: date_from
          in: query
          schema: { type: string, format: date-time }
        - name: date_to
          in: query
          schema: { type: string, format: date-time }
        - name: query
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 500 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: submitEvidence
      summary: Submit evidence
      description: Submits new evidence for a policy and/or assessment.
      tags: [Evidence]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvidenceSubmitRequest'
      responses:
        '201':
          description: Evidence submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evidence'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1.0/evidence/{evidence_id}:
    get:
      operationId: getEvidence
      summary: Get evidence
      description: Returns full evidence item with chain of custody and verification status.
      tags: [Evidence]
      parameters:
        - name: evidence_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Evidence details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evidence'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1.0/evidence/{evidence_id}/verify:
    post:
      operationId: verifyEvidence
      summary: Verify evidence
      description: Triggers integrity verification and chain-of-custody validation.
      tags: [Evidence]
      parameters:
        - name: evidence_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceVerification'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Evidence verification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /v1.0/evidence/export:
    post:
      operationId: exportEvidencePackage
      summary: Export evidence package
      description: Initiates async evidence package export.
      tags: [Evidence]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvidenceExportRequest'
      responses:
        '202':
          description: Export initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportPackage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1.0/evidence/export/{package_id}:
    get:
      operationId: getExportPackage
      summary: Get export package
      description: Returns export package status and download URL when complete.
      tags: [Evidence]
      parameters:
        - name: package_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Package status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportPackage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'API key in format: Bearer grc_live_... or Bearer grc_test_...'
    mtlsAuth:
      type: mutualTLS
      description: mTLS with SPIFFE SVID
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, detail, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
        trace_id:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
    Pagination:
      type: object
      properties:
        next_cursor:
          type: string
        has_next:
          type: boolean
        total:
          type: integer
...
    Evidence:
      type: object
      properties:
        evidence_id: { type: string }
        policy_id: { type: string }
        assessment_id: { type: string }
        source:
          type: object
          properties:
            type: { type: string }
            system: { type: string }
            collection_method: { type: string }
        evidence_type:
          type: string
          enum: [artifact, observation, interview, analysis, log]
        content:
          type: object
          properties:
            format: { type: string }
            data: { type: string }
            hash: { type: string }
        context:
          type: object
          properties:
            environment: { type: string }
            region: { type: string }
            timestamp:
              type: string
              format: date-time
            metadata:
              type: object
              additionalProperties: true
        validation:
          type: object
          properties:
            status: { type: string }
            validated_by: { type: string }
            validated_at:
              type: string
              format: date-time
            confidence_score: { type: number }
        verification_level:
          type: string
          enum: [L0, L1, L2, L3, L4]
        chain_of_custody:
          type: array
          items:
            type: object
            properties:
              action: { type: string }
              actor: { type: string }
              timestamp:
                type: string
                format: date-time
              hash: { type: string }
        retention_class: { type: string }
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    EvidenceSubmitRequest:
      type: object
      required: [source, evidence_type, content]
      properties:
        policy_id: { type: string }
        assessment_id: { type: string }
        source:
          type: object
          properties:
            type: { type: string }
            system: { type: string }
            collection_method: { type: string }
        evidence_type:
          type: string
          enum: [artifact, observation, interview, analysis, log]
        content:
          type: object
          properties:
            format: { type: string }
            data: { type: string }
        context:
          type: object
          properties:
            environment: { type: string }
            region: { type: string }
            metadata:
              type: object
              additionalProperties: true
        control_mapping:
          type: object
          properties:
            control_id: { type: string }
            framework: { type: string }
            control_title: { type: string }
    EvidenceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Evidence'
        pagination:
          $ref: '#/components/schemas/Pagination'
    EvidenceVerification:
      type: object
      properties:
        evidence_id: { type: string }
        verification_result:
          type: object
          properties:
            status: { type: string }
            verification_level:
              type: string
              enum: [L0, L1, L2, L3, L4]
            hash_match: { type: boolean }
            chain_of_custody_intact: { type: boolean }
            schema_valid: { type: boolean }
            control_mapping_valid: { type: boolean }
            verified_at:
              type: string
              format: date-time
            verified_by: { type: string }
    EvidenceExportRequest:
      type: object
      required: [framework, time_range, format]
      properties:
        framework: { type: string }
        time_range:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
        format:
          type: string
          enum: [json, pdf, csv, xml]
        include_chain_of_custody: { type: boolean }
    ExportPackage:
      type: object
      properties:
        package_id: { type: string }
        status:
          type: string
          enum: [processing, completed, failed]
        estimated_completion:
          type: string
          format: date-time
        download_url: { type: string }
        expires_at:
          type: string
          format: date-time
        package_hash: { type: string }
        manifest:
          type: object
          additionalProperties: true
...